+++ packages/expo-maps/ios/MapUtils.swift
@@ -24,3 +24,13 @@ func getLookAroundScene(from coordinate: CLLocationCoordinate2D) async throws ->
     throw SceneUnavailableAtLocationException()
   }
 }
+
+@available(iOS 17.0, *)
+extension View {
+  func renderCircle(_ circle: Circle) -> some MapContent {
+    let mapCircle = MapCircle(center: circle.clLocationCoordinate2D, radius: circle.radius)
+    return mapCircle
+      .stroke(circle.lineColor ?? .clear, lineWidth: circle.lineWidth ?? 0)
+      .foregroundStyle(circle.color)
+  }
+}